+++ src/api/exporter/index.ts
@@ -1,3 +1,4 @@
+import logger from "@/config/logger";
 import { KitaSchema } from "@/types/kitaschema";
 
 const ENV = process.env.APPLICATION_ENVIRONMENT;
@@ -28,7 +29,7 @@ const getItemsFromKey = <T>(key: string): Promise<T | null> => {
 const exportToJSON = (data: KitaSchema, fileName: string): Promise<KitaSchema | null> => {
   return new Promise((resolve, reject) => {
     try {
-      console.log("exporting data...");
+      logger.info("exporting...");
       data.Statistics.VideoStatistics.TotalVideos = data.UserItems.Videos.length ?? 0;
       data.Statistics.VideoStatistics.TotalDurationSeconds = data.Statistics.VideoStatistics.TotalDurationSeconds ?? 0;
       data.Statistics.TagStatistics.TotalTags = data.UserItems.Tags.length ?? 0;