+++ src/api/schemaGenerator/index.ts
@@ -7,6 +7,7 @@ const getKitaSchema = async (): Promise<KitaSchema> => {
   const videos = await IndexedDB.getAllVideos();
   const tags = await IndexedDB.getAllTags();
   const videoTagRelationships = await IndexedDB.getAllVideoTags();
+  const autoTags = await IndexedDB.getAllAutoTags();
   const isApplicationEnabled = await getItemsFromKey<boolean>(kitaSchema.ApplicationSettings.StorageKeys.ApplicationEnabledKey);
   const totalVideos = await getItemsFromKey<number>(
     kitaSchema.ApplicationSettings.StorageKeys.StatisticsKeys.VideoStatisticsKeys.TotalVideosKey
@@ -21,12 +22,14 @@ const getKitaSchema = async (): Promise<KitaSchema> => {
       Videos: videos ?? [],
       Tags: tags ?? [],
       VideoTagRelationships: videoTagRelationships ?? [],
+      AutoTags: autoTags ?? [],
     },
     ApplicationSettings: {
       IsReady: false, // @todo: implement
       IsApplicationEnabled: isApplicationEnabled ?? true,
       StorageKeys: {
         ApplicationEnabledKey: kitaSchema.ApplicationSettings.StorageKeys.ApplicationEnabledKey,
+        DefaultTagsInitializedKey: kitaSchema.ApplicationSettings.StorageKeys.DefaultTagsInitializedKey,
         VideoKey: kitaSchema.ApplicationSettings.StorageKeys.VideoKey,
         TagKey: kitaSchema.ApplicationSettings.StorageKeys.TagKey,
         ThemeKey: kitaSchema.ApplicationSettings.StorageKeys.ThemeKey,