+++ src/api/summaryStorage/tag.ts
@@ -1,12 +1,13 @@
 import { Callback } from "@/types/callback";
 import { kitaSchema } from "../videostorage";
+import logger from "@/config/logger";
 
 const TAG_TOTAL_KEY = kitaSchema.ApplicationSettings.StorageKeys.StatisticsKeys.TagStatisticsKeys.TotalTagsKey;
 const ENV = process.env.APPLICATION_ENVIRONMENT;
 
 const getTotalTagCount = (callback: Callback<number>) => {
   if (ENV === "dev") {
-    console.log("fetching tag total");
+    logger.info("fetching tag total");
     const items = localStorage.getItem(TAG_TOTAL_KEY);
     if (!items) {
       callback(0);
@@ -17,20 +18,20 @@ const getTotalTagCount = (callback: Callback<number>) => {
     return;
   }
   chrome.storage.local.get(TAG_TOTAL_KEY, (data) => {
-    console.log("fetching tag total");
+    logger.info("fetching tag total");
     const items = data?.[TAG_TOTAL_KEY] || 0;
     callback(items);
   });
 };
 
-const incrementTotalTags = () => {
+const incrementTotalTags = (value?: number) => {
   getTotalTagCount((count) => {
-    const newCount = count + 1;
+    const newCount = count + (value ?? 1);
     if (ENV === "dev") {
-      console.log("incrementing tag total");
+      logger.info("incrementing tag total");
       localStorage.setItem(TAG_TOTAL_KEY, JSON.stringify(newCount));
     } else {
-      console.log("incrementing tag total");
+      logger.info("incrementing tag total");
       chrome.storage.local.set({ [TAG_TOTAL_KEY]: newCount });
     }
   });
@@ -41,10 +42,10 @@ const decrementTotalTags = () => {
     if (count === 0) return;
     const newCount = count - 1;
     if (ENV === "dev") {
-      console.log("decrementing tag total");
+      logger.info("decrementing tag total");
       localStorage.setItem(TAG_TOTAL_KEY, JSON.stringify(newCount));
     } else {
-      console.log("decrementing tag total");
+      logger.info("decrementing tag total");
       chrome.storage.local.set({ [TAG_TOTAL_KEY]: newCount });
     }
   });
@@ -52,10 +53,10 @@ const decrementTotalTags = () => {
 
 const resetTotalTags = () => {
   if (ENV === "dev") {
-    console.log("resetting tag total");
+    logger.info("resetting tag total");
     localStorage.setItem(TAG_TOTAL_KEY, JSON.stringify(0));
   } else {
-    console.log("resetting tag total");
+    logger.info("resetting tag total");
     chrome.storage.local.set({ [TAG_TOTAL_KEY]: 0 });
   }
 };