+++ src/api/videostorage/index.ts
@@ -1,6 +1,7 @@
 import { Callback } from "@/types/callback";
 import { kitaSchema } from "../../data/kitaschema";
 import { IVideo } from "../../types/video";
+import logger from "../../config/logger";
 
 const VIDEO_KEY = kitaSchema.ApplicationSettings.StorageKeys.VideoKey;
 const ENV = process.env.APPLICATION_ENVIRONMENT;
@@ -12,7 +13,7 @@ const getVideoById = (id: string, videos: IVideo[]) => {
 
 const getVideos = (callback: Callback<IVideo[]>) => {
   if (ENV === "dev") {
-    console.log("fetching videos");
+    logger.info("fetching videos");
     const items = localStorage.getItem(VIDEO_KEY);
     if (!items) {
       callback([]);
@@ -23,7 +24,7 @@ const getVideos = (callback: Callback<IVideo[]>) => {
     return;
   }
   chrome.storage.local.get(VIDEO_KEY, (data) => {
-    console.log("fetching videos");
+    logger.info("fetching videos");
     const items = data?.[VIDEO_KEY] || [];
     callback(items);
   });
@@ -36,14 +37,14 @@ const setVideo = (video: IVideo, callback: Callback<IVideo>) => {
     localVideos.push(video);
 
     if (ENV === "dev") {
-      console.log("setting single video");
+      logger.info("setting single video");
       localStorage.setItem(VIDEO_KEY, JSON.stringify(localVideos));
       callback(video);
       return;
     }
 
     chrome.storage.local.set({ [VIDEO_KEY]: localVideos }, () => {
-      console.log("setting single video");
+      logger.info("setting single video");
       callback(video);
     });
   });
@@ -51,14 +52,14 @@ const setVideo = (video: IVideo, callback: Callback<IVideo>) => {
 
 const setVideos = (videos: IVideo[], callback: Callback<null>) => {
   if (ENV === "dev") {
-    console.log("setting videos");
+    logger.info("setting videos");
     localStorage.setItem(VIDEO_KEY, JSON.stringify(videos));
     callback(null);
     return;
   }
 
   chrome.storage.local.set({ [VIDEO_KEY]: videos }, () => {
-    console.log("setting videos");
+    logger.info("setting videos");
     callback(null);
   });
 };
@@ -72,14 +73,14 @@ const updateVideoById = (id: string, videoNext: IVideo, videos: IVideo[], callba
   });
 
   if (ENV === "dev") {
-    console.log("updating video id: ", id);
+    logger.info(`updating video id: ${id}`);
     localStorage.setItem(VIDEO_KEY, JSON.stringify(updatedVideos));
     callback(updatedVideos);
     return;
   }
 
   chrome.storage.local.set({ [VIDEO_KEY]: updatedVideos }, () => {
-    console.log("updating video id: ", id);
+    logger.info(`updating video id: ${id}`);
     callback(updatedVideos);
   });
 };
@@ -94,14 +95,14 @@ const deleteVideoById = (id: string, videos: IVideo[], callback: Callback<IVideo
   localVideos.splice(index, 1);
 
   if (ENV === "dev") {
-    console.log("delete video index: ", index);
+    logger.info(`delete video index: ${index}`);
     localStorage.setItem(VIDEO_KEY, JSON.stringify(localVideos));
     callback(localVideos);
     return;
   }
 
   chrome.storage.local.set({ [VIDEO_KEY]: localVideos }, () => {
-    console.log("delete video index: ", index);
+    logger.info(`delete video index: ${index}`);
     callback(localVideos);
   });
 };
@@ -109,14 +110,14 @@ const deleteVideoById = (id: string, videos: IVideo[], callback: Callback<IVideo
 // DELETE
 const deleteAllVideos = (callback: Callback<null>) => {
   if (ENV === "dev") {
-    console.log("deleting all videos");
+    logger.info("deleting all videos");
     localStorage.removeItem(VIDEO_KEY);
     callback(null);
     return;
   }
 
   chrome.storage.local.remove(VIDEO_KEY, () => {
-    console.log("deleting all videos");
+    logger.info("deleting all videos");
     callback(null);
   });
 };