+++ src/data/kitaschema.ts
@@ -5,12 +5,14 @@ export const kitaSchema: KitaSchema = {
     Videos: [],
     Tags: [],
     VideoTagRelationships: [],
+    AutoTags: [],
   },
   ApplicationSettings: {
     IsReady: false, // @todo: implement
     IsApplicationEnabled: true, // default enabled
     StorageKeys: {
       ApplicationEnabledKey: "kitamersion_application_enabled",
+      DefaultTagsInitializedKey: "kitamersion_default_tags_initialized",
       VideoKey: "kitamersion_video_logs",
       TagKey: "kitamersion_tag",
       ThemeKey: "kitamersion_theme",