+++ src/pages/background/index.ts
@@ -1,7 +1,15 @@
 import { getAnilistConfig, getAnilistAuthUrl, setAnilistAuth, setAnilistAuthStatus, setAnilistConfig } from "@/api/integration/anilist";
+import { incrementTotalTags } from "@/api/summaryStorage/tag";
+import { incrementTotalVideoDuration, incrementTotalVideos } from "@/api/summaryStorage/video";
+import { getDefaultTagsInitialized, setDefaultTagsInitialized } from "@/api/tags";
+import logger from "@/config/logger";
 import { INTEGRATION_ANILIST_AUTH_CONNECT, VIDEO_ADD } from "@/data/events";
 import IndexedDB from "@/db/index";
 import { AnilistAuth, AnilistConfig } from "@/types/integrations/anilist";
+import { IVideoTag } from "@/types/relationship";
+import { IVideo } from "@/types/video";
+import { generateUniqueCode } from "@/utils";
+import { v4 as uuidv4 } from "uuid";
 
 export type RuntimeResponse = {
   status: RuntimeStatus;
@@ -15,8 +23,8 @@ chrome.runtime.onMessage.addListener((request, sender, sendResponse) => {
   let parsedPayload;
   try {
     parsedPayload = JSON.parse(request.payload);
-  } catch (e) {
-    console.error("Error parsing payload", e);
+  } catch (error) {
+    logger.error(`Error parsing payload ${error}`);
     const response: RuntimeResponse = { status: "error", message: "error parsing payload" };
     sendResponse(response);
     return;
@@ -24,8 +32,47 @@ chrome.runtime.onMessage.addListener((request, sender, sendResponse) => {
 
   if (request.type === VIDEO_ADD) {
     (async () => {
-      console.log("Received ADD_VIDEO with payload:", parsedPayload);
-      await IndexedDB.addVideo(parsedPayload);
+      logger.info("received VIDEO_ADD event");
+      const { id, video_title, origin, video_duration } = parsedPayload as IVideo;
+      const uniqueCode = generateUniqueCode(video_title, origin);
+
+      try {
+        const hasExistingVideoItem = await IndexedDB.getVideoByUniqueCode(uniqueCode);
+        if (hasExistingVideoItem) {
+          logger.info("video already exists, skipping...");
+          return;
+        }
+
+        // apply auto tags
+        const autoTag = await IndexedDB.getAutoTagByOrigin(origin);
+        if (autoTag) {
+          parsedPayload.tags = autoTag.tags;
+        }
+
+        await IndexedDB.addVideo({ ...parsedPayload, unique_code: uniqueCode });
+        incrementTotalVideos();
+        incrementTotalVideoDuration(video_duration ?? 0);
+
+        if (autoTag) {
+          const videoTagRelationship: IVideoTag[] = autoTag.tags.map((tag_id) => {
+            return {
+              id: uuidv4(),
+              video_id: id,
+              tag_id: tag_id,
+            };
+          });
+
+          if (videoTagRelationship.length === 0) {
+            logger.warn("no video tag relationship to add");
+            return;
+          }
+          videoTagRelationship.forEach(async (videoTagRelationship) => {
+            await IndexedDB.addVideoTag(videoTagRelationship);
+          });
+        }
+      } catch (error) {
+        logger.error(`error while adding video: ${error}`);
+      }
     })();
     return;
   }
@@ -83,7 +130,7 @@ const authorizeAnilist = async (anilistConfig: AnilistConfig): Promise<boolean> 
 
     return true;
   } catch (error) {
-    console.error("Error authorizing anilist", error);
+    logger.error(`Error authorizing anilist ${error}`);
     return false;
   }
 };
@@ -99,3 +146,16 @@ const authorizeAnilist = async (anilistConfig: AnilistConfig): Promise<boolean> 
     }
   });
 })();
+
+chrome.runtime.onInstalled.addListener(() => {
+  (async () => {
+    // initialize default tags
+    getDefaultTagsInitialized(async (isInitialized) => {
+      if (!isInitialized) {
+        const addedTagsCount = await IndexedDB.initializeDefaultTags();
+        incrementTotalTags(addedTagsCount);
+        setDefaultTagsInitialized(() => {});
+      }
+    });
+  })();
+});