+++ src/pages/statistics/components/tagAssignedRadar.tsx
@@ -2,22 +2,26 @@ import LoadingState from "@/components/states/LoadingState";
 import { useTagContext } from "@/context/tagContext";
 import { useVideoContext } from "@/context/videoContext";
 import { Box } from "@chakra-ui/react";
-import React from "react";
+import React, { useMemo } from "react";
 import { Radar, RadarChart, PolarGrid, PolarAngleAxis, PolarRadiusAxis, ResponsiveContainer, Tooltip } from "recharts";
 
 const TagAssignedRadar = () => {
   const { isInitialized: videoContextInitialized, totalVideos } = useVideoContext();
   const { isInitialized: tagContextInitialized, tags } = useTagContext();
 
+  // Calculate the number of videos for each tag
+  const data = useMemo(
+    () =>
+      tags.map((tag) => ({
+        tagName: tag.name,
+        count: tag.id ? totalVideos.filter((video) => video.tags?.includes(tag.id as string)).length : 0,
+      })),
+    [tags, totalVideos]
+  );
+
   if (!videoContextInitialized && !tagContextInitialized) {
     return <LoadingState />;
   }
-
-  // Calculate the number of videos for each tag
-  const data = tags.map((tag) => ({
-    tagName: tag.name,
-    count: totalVideos.filter((video) => video.tags?.includes(tag.id)).length ?? 0,
-  }));
 
   return (
     <Box width={"full"} height={"500px"} boxShadow={"dark-lg"} rounded={"2xl"} p={4}>