+++ src/types/tag.ts
@@ -1,5 +1,5 @@
 export type ITag = {
-  id: string;
+  id?: string;
   name: string;
   code?: string;
   created_at?: number;